+++ evm/evm-codec/src/index.ts
@@ -1,12 +1,4 @@
-export { Src } from "./src";
-export { Sink } from "./sink";
-export type { Codec } from "./codec";
-export * from "./codecs/primitives";
-export * from "./contract-base";
-export type { EventParams } from "./abi-components/event";
-export type {
-  FunctionReturn,
-  FunctionArguments,
-} from "./abi-components/function";
-import keccak256 from "keccak256";
-export { keccak256 };
+export { Src } from './src'
+export { Sink } from './sink'
+export type { Codec, Struct, StructTypes } from './codec'
+export * from './codecs/primitives'