+++ solana/solana-processor/src/processor.ts
@@ -40,6 +40,10 @@ export interface RpcEndpointSettings {
      * Maximum number of requests in a single batch call
      */
     maxBatchCallSize?: number
+    /**
+     * HTTP headers
+     */
+    headers?: Record<string, string>
 }
 
 
@@ -365,6 +369,7 @@ export class SolanaBatchProcessor<F extends FieldSelection = {}> {
         }
         let client = new RpcClient({
             url: this.rpcEndpoint.url,
+            headers: this.rpcEndpoint.headers,
             maxBatchCallSize: this.rpcEndpoint.maxBatchCallSize ?? 100,
             requestTimeout:  this.rpcEndpoint.requestTimeout ?? 30_000,
             capacity: this.rpcEndpoint.capacity ?? 10,