+++ typeorm/typeorm-store/src/decorators/columns/StringColumn.ts
@@ -1,0 +1,15 @@
+import {Column} from './Column'
+import {ColumnCommonOptions} from './common'
+
+export type StringColumnOptions = Pick<
+    ColumnCommonOptions,
+    'name' | 'unique' | 'nullable' | 'default' | 'comment' | 'array'
+>
+
+/**
+ * StringColumn decorator is used to mark a specific class property as a `text` table column.
+ * Column value is transformed to `string` type.
+ */
+export function StringColumn(options?: StringColumnOptions): PropertyDecorator {
+    return Column('text', options)
+}