+++ util/rpc-client/src/index.ts
@@ -1,3 +1,4 @@
 export * from './errors'
 export * from './client'
 export {Subscription, SubscriptionHandle} from './subscriptions'
+export {HttpHeaders} from './interfaces'